Fix lack of sequence end
When an option was defined AFTER a sequence option, it would not be
properly parsed. E.g.:

  class Options(...) extends Sclip(...) {
    val seq = sopt[String](seq)
    val other = opt[String](other)
  }

with

  "--seq a b c --other d"

In this case, the sequence would pick up "--other" as an option.

This was because java.lang.String.startsWith does not match a pattern,
but instead a verbatim string.

Also add a test case to guarantee fix.